nook-weather/requirements.txt
requests
Flask
waitress
paste
geopy
pycairo
cairosvg
Pillow
